ver. 0.6.1 / db의 임시 데이터로 각 테이블에 저장.(예외처리 추가해야함.)

+++ src/main/java/kr/co/moviespring/web/repository/DirectorRepository.java
@@ -6,7 +6,7 @@ import kr.co.moviespring.web.entity.Director;
 
 @Mapper
 public interface DirectorRepository {
-    int save(Director director);
+    Long save(Director director);
     int update(Director director);
     int delete(Director director);
 }
